AI: Store best known continuations in transposition table.
Also refactors the negamax search. In particular, `NegamaxEvalReturnMove` is no longer necessary, as now we can fetch the move from the TT.

+++ AI/include/utils.h
@@ -210,8 +210,7 @@ std::string FileToStr(const std::string& file_name) {
   return res;
 }
 
-long long MillisSince(
-    const std::chrono::high_resolution_clock::time_point& start) {
+int MillisSince(const std::chrono::high_resolution_clock::time_point& start) {
   auto now = std::chrono::high_resolution_clock::now();
   return std::chrono::duration_cast<std::chrono::milliseconds>(now - start)
       .count();
